requirements.txt
aiohttp==3.6.2
appdirs==1.4.4
async-timeout==3.0.1
attrs==19.3.0
certifi==2020.6.20
chardet==3.0.4
compile==1.0.3
discord.py==1.3.4
distlib==0.3.1
filelock==3.0.12
flake8==3.8.3
idna==2.10
isort==5.0.9
jedi==0.17.1
mccabe==0.6.1
multidict==4.7.6
parso==0.7.0
pbr==5.4.5
pluggy==0.13.1
pycodestyle==2.6.0
pyflakes==2.2.0
python-dotenv==0.14.0
python-jsonrpc-server==0.3.4
python-language-server==0.34.1
requests==2.24.0
six==1.15.0
stevedore==3.1.0
urllib3==1.25.9
virtualenv==20.0.26
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
websockets==8.1
yapf==0.30.0
yarl==1.4.2
